requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dataclasses-json==0.5.13
Django==4.2.3
django-cors-headers==4.1.0
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==1.16.0
drf-yasg==1.21.7
feedparser==6.0.10
frozenlist==1.4.0
gnews==0.2.7
google==3.0.0
google-api-core==2.0.1
google-auth==2.0.2
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.1
googleapis-common-protos==1.53.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.1
langchain-community
langchain-openai
langchain
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
numexpr==2.8.4
numpy==1.25.0
openai
openapi-codec==1.3.2
openapi-schema-pydantic==1.2.4
packaging
pandas==2.0.3
protobuf==4.24.0
psycopg==3.1.9
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydub==0.25.1
PyJWT==2.8.0
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2023.3
PyYAML==6.0.1
requests==2.26.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
sgmllib3k==1.0.0
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post5
soupsieve==2.4.1
SQLAlchemy==2.0.19
sqlparse==0.4.4
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
spotipy==2.23.0
faiss-gpu==1.7.2
